rm'd migrations, made models abstract, cleaned admin
Removed migrations because they cause the tests to fail. Although the
errors in migrations can be mitigated when doing a regurlar
sycdb/runserver they do no work properly for the tests.

Made some models abstract, notably Person and Comany

Removed comments and instant messenger models and corresondingly made
adjustments to the admin.py for the changes to the models.